Crate ical

ical-rs 0.3.0

This is a library to parse the ICalendar format defined in RFC5545, as well as similar formats like VCard.

There are probably some issues to be taken care of, but the library should work for most cases. If you like to help out and would like to discuss any API changes, please contact me or create an issue.

The initial goal was to make a port from the ical.js library in JavaScript and many code/algorithms was taken from it but in order to but more 'Rusty' a complete rewrite have been made.

Documentation

Installing

Put this in your Cargo.toml:

[dependencies]
ical = "0.4.*"

There is several ways to use Ical depending on the level of parsing you want. Some new wrapper/formater could appeare in the next releases.

By default all the features are included but you can choose to include in you project only the needed ones.

Reexports

pub use parser::ical::IcalParser;

pub use parser::vcard::VcardParser;

pub use property::PropertyParser;

pub use line::LineReader;

Modules

line
Read and unfold a line from a BufRead.

parser
Wrapper around PropertyParser

property
Parse the result of LineReader into parts.